split out addwidget button
