+++ src/sentry/static/sentry/app/views/dashboardsV2/dashboard.tsx
@@ -1,28 +1,21 @@
 import React from 'react';
 import LazyLoad from 'react-lazyload';
 import {DndContext} from '@dnd-kit/core';
-import {
-  arrayMove,
-  rectSwappingStrategy,
-  SortableContext,
-  useSortable,
-} from '@dnd-kit/sortable';
+import {arrayMove, rectSwappingStrategy, SortableContext} from '@dnd-kit/sortable';
 import styled from '@emotion/styled';
 
 import {openAddDashboardWidgetModal} from 'app/actionCreators/modal';
 import {loadOrganizationTags} from 'app/actionCreators/tags';
 import {Client} from 'app/api';
-import {IconAdd} from 'app/icons';
 import space from 'app/styles/space';
 import {GlobalSelection, Organization} from 'app/types';
 import withApi from 'app/utils/withApi';
 import withGlobalSelection from 'app/utils/withGlobalSelection';
 
+import AddWidget, {ADD_WIDGET_BUTTON_DRAG_ID} from './addWidget';
 import SortableWidget from './sortableWidget';
-import {WidgetWrapper} from './styles';
 import {DashboardDetails, Widget} from './types';
 
-const ADD_WIDGET_BUTTON_DRAG_ID = 'add-widget-button';
 type Props = {
   api: Client;
   organization: Organization;
@@ -173,60 +166,6 @@ const WidgetContainer = styled('div')`
     grid-template-columns: 1fr;
   }
 `;
-
-const AddWidgetWrapper = styled('a')`
-  width: 100%;
-  height: 110px;
-  border: 2px dashed ${p => p.theme.border};
-  border-radius: ${p => p.theme.borderRadius};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-function AddWidget(props: {onClick: () => void}) {
-  const {onClick} = props;
-
-  const initialStyles = {
-    x: 0,
-    y: 0,
-    scaleX: 1,
-    scaleY: 1,
-  };
-
-  const {setNodeRef, transform} = useSortable({
-    disabled: true,
-    id: ADD_WIDGET_BUTTON_DRAG_ID,
-    transition: null,
-  });
-
-  return (
-    <WidgetWrapper
-      key="add"
-      ref={setNodeRef}
-      displayType="big_number"
-      layoutId={ADD_WIDGET_BUTTON_DRAG_ID}
-      style={{originX: 0, originY: 0}}
-      animate={
-        transform
-          ? {
-              x: transform.x,
-              y: transform.y,
-              scaleX: transform?.scaleX && transform.scaleX <= 1 ? transform.scaleX : 1,
-              scaleY: transform?.scaleY && transform.scaleY <= 1 ? transform.scaleY : 1,
-            }
-          : initialStyles
-      }
-      transition={{
-        duration: 0.25,
-      }}
-    >
-      <AddWidgetWrapper key="add" data-test-id="widget-add" onClick={onClick}>
-        <IconAdd size="lg" isCircled color="inactive" />
-      </AddWidgetWrapper>
-    </WidgetWrapper>
-  );
-}
 
 function generateWidgetId(widget: Widget, index: number) {
   return widget.id ?? `index-${index}`;
